Move the order object functionality into a custom service
This will allow later use during order completion.
