+++ web/modules/custom/ilr_registrations/ilr_registrations.services.yml
@@ -2,9 +2,12 @@ services:
   ilr_registrations_commerce_event_subscriber:
     class: Drupal\ilr_registrations\EventSubscriber\CommerceEventSubscriber
     tags:
-    - { name: event_subscriber }
+      - { name: event_subscriber }
   ilr_registrations_salesforce_event_subscriber:
     class: Drupal\ilr_registrations\EventSubscriber\SalesforceEventSubscriber
     arguments: []
     tags:
       - { name: event_subscriber }
+  ilr_registrations.serialized_order:
+    class: Drupal\ilr_registrations\SerializedOrderManager
+    arguments: ['@entity.manager']
